fix(utils): fix linting issues
