+++ convert_source_description/tests/test_convert_source_description.py
@@ -1,7 +1,7 @@
 """Tests for convert_source_description.py."""
 
-import pytest
 from unittest.mock import patch
+import pytest
 
 from convert_source_description import convert_source_description, main
 
@@ -27,9 +27,7 @@ class TestConvertSourceDescription:
                 "convert_source_description.TextcriticsUtils.create_textcritics",
                 return_value=textcritics,
             ) as mock_textcritics,
-            patch(
-                "convert_source_description.FileUtils.write_json"
-            ) as mock_write,
+            patch("convert_source_description.FileUtils.write_json") as mock_write,
         ):
             convert_source_description(file_path)
             return mock_sources, mock_textcritics, mock_write
@@ -77,9 +75,7 @@ class TestConvertSourceDescription:
                 "convert_source_description.TextcriticsUtils.create_textcritics",
                 return_value={},
             ),
-            patch(
-                "convert_source_description.FileUtils.write_json"
-            ) as mock_write,
+            patch("convert_source_description.FileUtils.write_json") as mock_write,
         ):
             convert_source_description("path/to/source.docx")
         written_data = [c.args[0] for c in mock_write.call_args_list]
@@ -117,17 +113,13 @@ class TestMain:
     def test_calls_convert_source_description_with_parsed_args(self, monkeypatch):
         """Test that main() passes the file_path arg to convert_source_description."""
         monkeypatch.setattr("sys.argv", ["prog", "my/dir/myfile.docx"])
-        with patch(
-            "convert_source_description.convert_source_description"
-        ) as mock_fn:
+        with patch("convert_source_description.convert_source_description") as mock_fn:
             main()
         mock_fn.assert_called_once_with("my/dir/myfile.docx")
 
     def test_file_path_argument_is_passed_correctly(self, monkeypatch):
         """Test that the file_path positional argument is forwarded."""
         monkeypatch.setattr("sys.argv", ["prog", "some/path/my_source.docx"])
-        with patch(
-            "convert_source_description.convert_source_description"
-        ) as mock_fn:
+        with patch("convert_source_description.convert_source_description") as mock_fn:
             main()
         assert mock_fn.call_args.args[0] == "some/path/my_source.docx"
